backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
pydantic==2.7.4
python-dotenv==1.0.1